Fixes hybrid mana symbols
The hybrid symbols did not like having the "pull-right" class applied to
them due to their absolute positioning.  I've overridden the positioning
with relative to fix this issue.

I have also added the mana symbols in the decklist to belong to a span
and make the span pulled to the right instead of each symbol.